move repr to base class
