+++ skore/src/skore/_sklearn/_base.py
@@ -10,6 +10,7 @@ from skore._project.git import git_commit
 from skore._sklearn._checks._utils import CheckNotApplicable
 from skore._sklearn._checks.base import Check, CheckCode
 from skore._sklearn._checks.model_checks import _BUILTIN_CHECKS
+from skore._sklearn.types import DataSource
 from skore._utils._progress_bar import track
 from skore._utils.repr.base import (
     AccessorHelpMixin,
@@ -18,6 +19,8 @@ from skore._utils.repr.base import (
 )
 
 if TYPE_CHECKING:
+    import pandas as pd
+
     from skore._sklearn._checks.accessor import _ChecksAccessor
 
 
@@ -193,3 +196,30 @@ class BaseMetricsAccessor(_BaseAccessor, Generic[ParentT]):
     def __dir__(self) -> list[str]:
         """Add custom metrics to __dir__ for tab-completion."""
         return list(set(super().__dir__()).union(self.available()))
+
+    def _formatted_summary_frame(
+        self,
+        *,
+        data_source: DataSource = "test",
+        metric: str | list[str] | None = None,
+    ) -> pd.DataFrame:
+        """Metric summary.
+
+        Used for displaying the report.
+        """
+        return self.summarize().frame()
+
+    def __repr__(self) -> str:
+        return (
+            "Metrics summary:\n"
+            f"{self._formatted_summary_frame()!r}\n"
+            "Explore available methods with .help()."
+        )
+
+    def _repr_html_(self) -> str:
+        return (
+            "<p>Metrics summary:</p>"
+            f"{self._formatted_summary_frame()._repr_html_()}"
+            '<p role="note">Explore available methods with '
+            "<code>.help()</code>.</p>"
+        )
